Support paginated methods in async client (#1615)
* Support paginated methods in async client

To be able to support both sync and async in paginated methods, we need
to push the pagination logic to the client so that the coroutine
handling can be handled there and the top level method retains the lower
level returned type. This refactors it a bit to be able to do that.

* Less duplication

* Remove some more duplication

* Remove specific method

* Don't pass parameters that breaks data handling e3d89d96712d0bc447fd2c847dbebd4ef3dd6d67